+++ src/main/java/funoform/mdp/gui/Gui.java
@@ -97,7 +97,6 @@ public class Gui extends JPanel implements RaiseWindowRequestListener {
 		mDirSelector = new DirectoryPicker(mCtrl, new PathSelectionListener() {
 			@Override
 			public void setPathSelected(Path selPath, boolean isRecursive) {
-				mCurBrowsingDir = selPath;
 				ctrl.playDir(selPath, isRecursive);
 				populateDirSongList();
 
@@ -274,6 +273,13 @@ public class Gui extends JPanel implements RaiseWindowRequestListener {
 				SwingUtilities.invokeLater(new Runnable() {
 					@Override
 					public void run() {
+
+						// if the directory being played changed since last time, then we need to get an
+						// updated list of queued songs
+						if (mCurBrowsingDir != settings.playingDir) {
+							mCurBrowsingDir = settings.playingDir;
+							populateDirSongList();
+						}
 
 						// if the song has changed since last time, consider scrolling the song list to
 						// the new song